Add check for response.usage before token limit
This was reported in jekalmin/extended_openai_conversation#273, but then closed as the author found a workaround. The workaround is not the correct solution though, because that involves bypassing Open WebUI and using the `/ollama/` api directly. Doing so removes the ability to use Open WebUI's custom models.

The correct fix is to check that `.usage` exists before using it.

Note that I left the later `reason="length"` usage, because that error type does imply that usage data is accessible, so if you get to that point, then there really should be a `.usage` object available.